requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.2.3
flask-babel==3.1.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.2
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imgaug==0.4.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
linkify-it-py==2.0.0
lmdb==1.4.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.2
onnx==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.1
paddle-bfloat==0.1.7

paddleocr>= 2.0.1
paddlepaddle==2.5.1

PyMuPDF==1.21.1

pandas==2.0.0
pdf2docx==0.5.6
Pillow==9.5.0
pluggy==1.0.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.5
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9

pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
pytweening==1.0.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
rarfile==4.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.10.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
sympy==1.11.1
termcolor==2.2.0
tifffile==2023.4.12
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
tritonclient==2.32.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
visualdl==2.5.1
websockets==11.0.2
Werkzeug==2.2.3
x2paddle==1.4.1
yarl==1.9.1
zope.event==4.6
zope.interface==6.0